restructured packages for clarity.
Also, when users click next or back when signing in, their input is maintained until they click sign in.
