+++ src/main/Java/interface_adapter/choose_account_type/AccountTypePresenter.java
@@ -3,7 +3,7 @@ package interface_adapter.choose_account_type;
 import interface_adapter.ViewManagerModel;
 import interface_adapter.event_poster_signup.EventPosterSignupViewModel;
 import interface_adapter.general_user_signup.UserSignupViewModel;
-import use_case.choose_account_type.AccountTypeOutputBoundary;
+import use_case.signup.choose_account_type.AccountTypeOutputBoundary;
 
 public class AccountTypePresenter implements AccountTypeOutputBoundary {
