--- src/main/Java/use_case/Tempfile.java
@@ -1,4 +1,0 @@
-package use_case;
-
-public class Tempfile {
-}
