--- src/main/Java/use_case/choose_account_type/AccountTypeOutputData.java
@@ -1,4 +1,0 @@
-package use_case.choose_account_type;
-
-public class AccountTypeOutputData {
-}
