+++ src/main/Java/use_case/signup/choose_account_type/AccountTypeOutputData.java
@@ -1,0 +1,4 @@
+package use_case.signup.choose_account_type;
+
+public class AccountTypeOutputData {
+}
